requirements.txt
streamlit
gtts
openai-whisper
streamlit-mic-recorder
asgiref==3.3.4
diff-match-patch==20200713
ffmpeg
pytz==2021.1
typing-extensions==3.10.0.0